requirement.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
contourpy==1.1.0
cycler==0.11.0
dacite==1.8.1
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
easydict==1.11
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.104.0
FastAPI-SQLAlchemy==0.2.1
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==1.1.0
grpcio==1.57.0
gspread==5.12.0
h11==0.14.0
h5py==3.9.0
htmlmin==0.1.12
httplib2==0.22.0
hydra-core==1.3.2
idna==2.10
ImageHash==4.3.1
imageio==2.31.1
importlib-metadata @ file:///C:/b/abs_20ndzb2j6v/croot/importlib-metadata_1678997085534/work
importlib-resources==6.0.1
ipykernel @ file:///C:/b/abs_07rkft_vaz/croot/ipykernel_1691121700587/work
ipython @ file:///C:/b/abs_254uk73z5b/croot/ipython_1691532131313/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client @ file:///C:/b/abs_059idvdagk/croot/jupyter_client_1680171872444/work
jupyter_core @ file:///C:/b/abs_9d0ttho3bs/croot/jupyter_core_1679906581955/work
jws==0.1.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mlxtend==0.22.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multimethod==1.9.1
nbformat==5.9.2
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
networkx==3.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==2.0.3
pandas-profiling==3.6.6
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.3
phik==0.12.3
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.0
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
plotly==5.16.0
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
psycopg2-binary==2.9.9
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycryptodome==3.19.0
pydantic==1.10.12
pydot==1.4.2
PyGetWindow==0.0.9
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
Pyrebase==3.0.27
Pyrebase4==4.7.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-jwt==2.0.1
pytweening==1.0.7
pytz==2023.3
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
referencing==0.30.2
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.8.12
torch==2.0.1
torchvision==0.15.2
tornado @ file:///C:/b/abs_61jhmrrua1/croot/tornado_1690848767317/work
tqdm==4.66.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.1
unicorn==2.0.1.post1
urllib3==1.25.11
uvicorn==0.23.2
visions==0.7.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.3.6
wikipedia==1.4.0
wordcloud==1.9.2
wrapt==1.15.0
ydata-profiling==4.5.1
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
